Move calnetdeposit and calremaining to report, add tip for dump data
